requirements.txt
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==2.0.0
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
fonttools==4.61.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
kiwisolver==1.4.9
logzero==1.7.0
matplotlib==3.10.8
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pycparser==2.23
pycryptodome==3.23.0
pyotp==2.8.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2023.3
rdflib==6.3.2
rdflib-jsonld==0.6.2
requests==2.31.0
setuptools==80.9.0
simplejson==3.19.1
six==1.16.0
smartapi-python==1.4.8
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2025.3
urllib3==2.0.3
websocket-client==1.8.0
zope.interface==8.1.1
